Add Eagle View dashboard and queries
Introduce the 'Eagle View' Metabase dashboard (dashboards/metabase/06_eagle_view.json) to provide a platform-wide overview of locations, finances, harvest, environment, attestations, sensors, and monthly trends. Add corresponding SQL query files (sql/06_eagle_view_overview.sql through sql/12_eagle_view_monthly_trend.sql). Update README to document the new dashboard and clarify that queries use the public schema / bare table names. Also apply minor SQL fixes to ensure numeric casting for rounding in sql/02_crop_noi.sql and sql/05_loss_rate.sql.

+++ dashboards/metabase/sql/02_crop_noi.sql
@@ -14,7 +14,7 @@ SELECT
     WHEN SUM(se.total_amount) - SUM(se.return_amount + se.discount_amount) > 0 
     THEN ROUND(
       (SUM(se.total_amount) - SUM(se.return_amount + se.discount_amount) - SUM(ee.amount)) / 
-      NULLIF(SUM(se.total_amount) - SUM(se.return_amount + se.discount_amount), 0) * 100, 
+      NULLIF(SUM(se.total_amount) - SUM(se.return_amount + se.discount_amount), 0) * 100::numeric, 
       2
     )
     ELSE 0
